+++ CHANGELOG.md
@@ -2,9 +2,12 @@ InfrastructureModels.jl Change Log
 ==================================
 
 ### Staged
+- nothing
+
+### v0.0.3
 - Added compare_dict function
 - Added arrays_to_dicts! function
-- Fixed bug when matlab function returns a value that is not called "mpc"
+- Fixed bug when a matlab function returns a value that is not called "mpc"
 
 ### v0.0.2
 - Added dict summary function